Fix source product generation and enable generation of background/rms map cutouts (#445)
* Initial debugging

* Tentative (messy) fixes

* Added handling for background images

* Added initial rms/bkg data fetching

* Added handling for background and rms cutout generation

* Fixed test_image_measure_coord_pixel_values_rms

* Updated _save_noisemap_cutout docstring

* Added get_bkg_img() tests

* Added rms and bkg cutout options to find_sources

* Fixed syntax error in save_noisemap_cutout

* fixed testing

* Fixed bkg path handling

* Updated find_sources CLI help

* Update find_sources output docs

* Updated examples

+++ vasttools/query.py
@@ -29,6 +29,7 @@ from astropy.coordinates import SkyCoord
 from astropy.io import fits
 from astropy.wcs import WCS
 from astropy.utils.exceptions import AstropyWarning, AstropyDeprecationWarning
+from astropy.nddata.utils import NoOverlapError
 
 from functools import partial
 
@@ -240,6 +241,7 @@ class Query:
                 )
             )
         self.ncpu = ncpu
+        self.logger.debug(f"Using {self.ncpu} CPUs")
 
         if coords is None and len(source_names) == 0 and len(planets) == 0:
             raise QueryInitError(
@@ -430,6 +432,9 @@ class Query:
         Returns:
             `True` if settings are acceptable, `False` otherwise.
         """
+        
+        self.logger.debug("Using settings: ")
+        self.logger.debug(self.settings)
 
         if self.settings['tiles'] and self.settings['stokes'].lower() != "i":
             if self.vast_full:
@@ -541,12 +546,20 @@ class Query:
 
         return all_available
 
-    def _get_all_cutout_data(self, imsize: Angle) -> pd.DataFrame:
+    def _get_all_cutout_data(self,
+                             imsize: Angle,
+                             img: bool = True,
+                             rms: bool = False,
+                             bkg: bool = False
+                             ) -> pd.DataFrame:
         """
         Get cutout data and selavy components for all sources.
 
         Args:
             imsize: Size of the requested cutout.
+            img: Fetch image data, defaults to `True`.
+            rms: Fetch rms data, defaults to `False`.
+            bkg: Fetch bkg data, defaults to `False`.
 
         Returns:
             Dataframe containing the cutout data of all measurements in
@@ -573,8 +586,14 @@ class Query:
             'header': 'O',
             'selavy_overlay': 'O',
             'beam': 'O',
+            'rms_data': 'O',
+            'rms_wcs': 'O',
+            'rms_header': 'O',
+            'bkg_data': 'O',
+            'bkg_wcs': 'O',
+            'bkg_header': 'O',
             'name': 'U',
-            'dateobs': 'datetime64[ns]'
+            'dateobs': 'datetime64[ns]',
         }
 
         cutouts = (
@@ -584,6 +603,9 @@ class Query:
                 self._grouped_fetch_cutouts,
                 imsize=imsize,
                 meta=meta,
+                img=img,
+                rms=rms,
+                bkg=bkg,
             ).compute(num_workers=self.ncpu,
                       scheduler=self.settings['scheduler']
                       )
@@ -598,6 +620,8 @@ class Query:
     def _gen_all_source_products(
         self,
         fits: bool = True,
+        rms: bool = False,
+        bkg: bool = False,
         png: bool = False,
         ann: bool = False,
         reg: bool = False,
@@ -638,6 +662,10 @@ class Query:
 
         Args:
             fits: Create and save fits cutouts, defaults to `True`.
+            rms: Create and save fits cutouts of the rms images,
+                defaults to `True`.
+            bkg: Create and save fits cutouts of the bkg images,
+                defaults to `True`.
             png: Create and save png postagestamps, defaults to `False`.
             ann: Create and save kvis annotation files for all components,
                 defaults to `False`
@@ -703,11 +731,15 @@ class Query:
                 ' used.'
             )
 
-        if sum([fits, png, ann, reg]) > 0:
+        if sum([fits, rms, bkg, png, ann, reg]) > 0:
             self.logger.info(
                 "Fetching cutout data for sources..."
             )
-            cutouts_df = self._get_all_cutout_data(imsize)
+            cutouts_df = self._get_all_cutout_data(imsize,
+                                                   img=fits,
+                                                   rms=rms,
+                                                   bkg=bkg,
+                                                   )
             self.logger.info('Done.')
             if cutouts_df.empty:
                 fits = False
@@ -738,6 +770,8 @@ class Query:
         produce_source_products_multi = partial(
             self._produce_source_products,
             fits=fits,
+            rms=rms,
+            bkg=bkg,
             png=png,
             ann=ann,
             reg=reg,
@@ -772,45 +806,52 @@ class Query:
             plot_dpi=plot_dpi
         )
 
-        original_sigint_handler = signal.signal(
-            signal.SIGINT, signal.SIG_IGN
-        )
-        signal.signal(signal.SIGINT, original_sigint_handler)
-        workers = Pool(processes=self.ncpu, maxtasksperchild=100)
+        if self.settings['scheduler'] == 'processes':
+            original_sigint_handler = signal.signal(
+                signal.SIGINT, signal.SIG_IGN
+            )
+            signal.signal(signal.SIGINT, original_sigint_handler)
+            workers = Pool(processes=self.ncpu, maxtasksperchild=100)
 
-        try:
-            workers.map(
-                produce_source_products_multi,
-                to_process,
-            )
-        except KeyboardInterrupt:
-            self.logger.error(
-                "Caught KeyboardInterrupt, terminating workers."
-            )
-            workers.terminate()
-            sys.exit()
-        except Exception as e:
-            self.logger.error(
-                "Encountered error!."
-            )
-            self.logger.error(
-                e
-            )
-            workers.terminate()
-            sys.exit()
-        finally:
-            self.logger.debug("Closing workers.")
-            # Using terminate below as close was prone to hanging
-            # when join is called. I believe the hang comes from
-            # a child processes not getting returned properly
-            # because of the large number of file I/O.
-            workers.terminate()
-            workers.join()
+            try:
+                workers.map(
+                    produce_source_products_multi,
+                    to_process,
+                )
+            except KeyboardInterrupt:
+                self.logger.error(
+                    "Caught KeyboardInterrupt, terminating workers."
+                )
+                workers.terminate()
+                sys.exit()
+            except Exception as e:
+                self.logger.error(
+                    "Encountered error!."
+                )
+                self.logger.error(
+                    e
+                )
+                workers.terminate()
+                sys.exit()
+            finally:
+                self.logger.debug("Closing workers.")
+                # Using terminate below as close was prone to hanging
+                # when join is called. I believe the hang comes from
+                # a child processes not getting returned properly
+                # because of the large number of file I/O.
+                workers.terminate()
+                workers.join()
+        elif self.settings['scheduler'] == 'single-threaded' or self.ncpu == 1:
+            for result in map(produce_source_products_multi, to_process):
+                pass
+            
 
     def _produce_source_products(
         self,
         i: Tuple[Source, pd.DataFrame],
         fits: bool = True,
+        rms: bool = False,
+        bkg: bool = False,
         png: bool = False,
         ann: bool = False,
         reg: bool = False,
@@ -850,6 +891,10 @@ class Query:
         Args:
             i: Tuple containing source and cutout data.
             fits: Create and save fits cutouts, defaults to `True`.
+            rms: Create and save fits cutouts of the rms images,
+                defaults to `True`.
+            bkg: Create and save fits cutouts of the bkg images,
+                defaults to `True`.
             png: Create and save png postagestamps, defaults to `False`.
             ann: Create and save kvis annotation files for all components,
                 defaults to `False`.
@@ -910,9 +955,13 @@ class Query:
         """
 
         source, cutout_data = i
+        
+        self.logger.debug(f"Producing source products for {source.name}")
 
         if fits:
             source.save_all_fits_cutouts(cutout_data=cutout_data)
+        if sum([rms, bkg]) > 1:
+            source._save_all_noisemap_cutouts(cutout_data, rms=rms, bkg=bkg)
 
         if png:
             source.save_all_png_cutouts(
@@ -1021,7 +1070,12 @@ class Query:
         return s
 
     def _grouped_fetch_cutouts(
-        self, group: pd.DataFrame, imsize: Angle
+        self,
+        group: pd.DataFrame,
+        imsize: Angle,
+        img: bool = True,
+        rms: bool = False,
+        bkg: bool = False
     ) -> pd.DataFrame:
         """
         Function that handles fetching the cutout data per
@@ -1031,11 +1085,15 @@ class Query:
         Args:
             group: Catalogue of sources grouped by field.
             imsize: Size of the requested cutout.
+            img: Fetch image data, defaults to `True`.
+            rms: Fetch rms data, defaults to `False`.
+            bkg: Fetch bkg data, defaults to `False`.
 
         Returns:
             Dataframe containing the cutout data for the group.
         """
         image_file = group.iloc[0]['image']
+        self.logger.debug(f"Fetching cutouts from {image_file}")
 
         try:
             image = Image(
@@ -1047,27 +1105,100 @@ class Query:
                 tiles=self.settings['tiles'],
                 corrected_data=self.corrected_data
             )
-
-            image.get_img_data()
-
-            cutout_data = group.apply(
-                self._get_cutout,
-                args=(image, imsize),
-                axis=1,
-                result_type='expand'
-            ).rename(columns={
-                0: "data",
-                1: "wcs",
-                2: "header",
-                3: "selavy_overlay",
-                4: "beam"
-            })
+            
+            if img:
+                image.get_img_data()
+                img_cutout_data = group.apply(
+                    self._get_cutout,
+                    args=(image, imsize),
+                    axis=1,
+                    result_type='expand'
+                ).rename(columns={
+                    0: "data",
+                    1: "wcs",
+                    2: "header",
+                    3: "selavy_overlay",
+                    4: "beam"
+                })
+            else:
+                img_cutout_data = pd.DataFrame([[None]*5]*len(group),
+                    columns=[
+                        'data',
+                        'wcs',
+                        'header',
+                        'selavy_overlay',
+                        'beam',
+                    ]
+                )
+            if rms:
+                image.get_rms_img()
+                rms_cutout_data = group.apply(
+                    self._get_cutout,
+                    args=(image, imsize),
+                    img=False,
+                    rms=True,
+                    axis=1,
+                    result_type='expand'
+                ).rename(columns={
+                    0: "rms_data",
+                    1: "rms_wcs",
+                    2: "rms_header",
+                }).drop(columns=[3, 4])
+            else:
+                rms_cutout_data = pd.DataFrame([[None]*3]*len(group),
+                    columns=[
+                        'rms_data',
+                        'rms_wcs',
+                        'rms_header',
+                    ]
+                )
+                
+            if bkg:
+                image.get_bkg_img()
+                bkg_cutout_data = group.apply(
+                    self._get_cutout,
+                    args=(image, imsize),
+                    img=False,
+                    bkg=True,
+                    axis=1,
+                    result_type='expand'
+                ).rename(columns={
+                    0: "bkg_data",
+                    1: "bkg_wcs",
+                    2: "bkg_header",
+                }).drop(columns=[3, 4])
+            else:
+                bkg_cutout_data = pd.DataFrame([[None]*3]*len(group),
+                    columns=[
+                        'bkg_data',
+                        'bkg_wcs',
+                        'bkg_header',
+                    ]
+                )
+                
+            self.logger.debug("Generated all cutout data")
+            
+            to_concat = [img_cutout_data, rms_cutout_data, bkg_cutout_data]
+            cutout_data = pd.concat(to_concat, axis=1).dropna(how='all')
+            
+            self.logger.debug("Concatenated into cutout_data")
+            
+            if bkg_cutout_data['bkg_data'].values == rms_cutout_data['rms_data'].values:
+                self.logger.warning("Background and RMS data are identical!")
+            
+            self.logger.debug(cutout_data.columns)
+            self.logger.debug(len(cutout_data))
+            self.logger.debug(group['name'].values)
 
             cutout_data['name'] = group['name'].values
+            self.logger.debug(cutout_data['name'])
             cutout_data['dateobs'] = group['dateobs'].values
+            self.logger.debug(cutout_data['dateobs'])
 
             del image
         except Exception as e:
+            self.logger.warning("Caught exception inside _grouped_fetch_cutouts")
+            self.logger.warning(e)
             cutout_data = pd.DataFrame(columns=[
                 'data',
                 'wcs',
@@ -1075,14 +1206,28 @@ class Query:
                 'selavy_overlay',
                 'beam',
                 'name',
-                'dateobs'
+                'dateobs',
+                'rms_data',
+                'rms_wcs',
+                'rms_header',
+                'bkg_data',
+                'bkg_wcs',
+                'bkg_header',
             ])
+        
+        # Drop the cutouts that raised a NoOverlapError
+        cutout_data.dropna(inplace=True)
 
         return cutout_data
 
     def _get_cutout(
-        self, row: pd.Series, image: Image,
-        size: Angle = Angle(5. * u.arcmin)
+        self,
+        row: pd.Series,
+        image: Image,
+        size: Angle = Angle(5. * u.arcmin),
+        img: bool = True,
+        rms: bool = False,
+        bkg: bool = False
     ) -> Tuple[pd.DataFrame, WCS, fits.Header, pd.DataFrame, Beam]:
         """
         Create cutout centered on a source location
@@ -1092,50 +1237,86 @@ class Query:
                 interest
             image: Image to create cutout from.
             size: Size of the cutout, defaults to Angle(5.*u.arcmin).
+            img: Make a cutout from the image data, defaults to `True`.
+            rms: Make a cutout from the rms data, defaults to `False`.
+            bkg: Make a cutout from the bkg data, defaults to `False`.
 
         Returns:
             Tuple containing cutout data, WCS, image header, associated
             selavy components and beam information.
+        
+        Raises:
+            ValueError: Exactly one of img, rms or bkg must be `True`
         """
+        
+        if sum([img,rms,bkg]) != 1:
+            raise ValueError("Exactly one of img, rms or bkg must be True")
 
-        cutout = Cutout2D(
-            image.data,
-            position=row.skycoord,
-            size=size,
-            wcs=image.wcs
-        )
+        if img:
+            thedata = image.data
+            thewcs = image.wcs
+            theheader = image.header.copy()
+            thepath = image.imgpath
+        elif rms:
+            thedata = image.rms_data
+            thewcs = image.rms_wcs
+            theheader = image.rms_header.copy()
+            thepath = image.rmspath
+        elif bkg:
+            thedata = image.bkg_data
+            thewcs = image.bkg_wcs
+            theheader = image.bkg_header.copy()
+            thepath = image.bkgpath
+        
+        self.logger.debug(f"Using data from {thepath}")
 
-        selavy_components = read_selavy(row.selavy, cols=[
-            'island_id',
-            'ra_deg_cont',
-            'dec_deg_cont',
-            'maj_axis',
-            'min_axis',
-            'pos_ang'
-        ])
+        try:
+            cutout = Cutout2D(
+                thedata,
+                position=row.skycoord,
+                size=size,
+                wcs=thewcs
+            )
+        except NoOverlapError:
+            self.logger.warning(f"Unable to create cutout for {row['name']}.")
+            self.logger.warning(f"Image path: {thepath}")
+            self.logger.warning(f"Coordinate: {row.skycoord.to_string()}")
+            return (None, None, None, None, None)
 
-        selavy_coords = SkyCoord(
-            selavy_components.ra_deg_cont.values,
-            selavy_components.dec_deg_cont.values,
-            unit=(u.deg, u.deg)
-        )
+        if img:
+            selavy_components = read_selavy(row.selavy, cols=[
+                'island_id',
+                'ra_deg_cont',
+                'dec_deg_cont',
+                'maj_axis',
+                'min_axis',
+                'pos_ang'
+            ])
 
-        selavy_components = filter_selavy_components(
-            selavy_components,
-            selavy_coords,
-            size,
-            row.skycoord
-        )
+            selavy_coords = SkyCoord(
+                selavy_components.ra_deg_cont.values,
+                selavy_components.dec_deg_cont.values,
+                unit=(u.deg, u.deg)
+            )
 
-        header = image.header.copy()
-        header.update(cutout.wcs.to_header())
+            selavy_components = filter_selavy_components(
+                selavy_components,
+                selavy_coords,
+                size,
+                row.skycoord
+            )
+            
+            del selavy_coords
+            
+            beam = image.beam
+        else:
+            beam = None
+            selavy_components = None
 
-        beam = image.beam
-
-        del selavy_coords
+        theheader.update(cutout.wcs.to_header())
 
         return (
-            cutout.data, cutout.wcs, header, selavy_components, beam
+            cutout.data, cutout.wcs, theheader, selavy_components, beam
         )
 
     def find_sources(self) -> None:
